[graph_trainer] Extract registry.py for pass/hook registries
Move MEMORY_POLICY_REGISTRY, PASS_PIPELINE_REGISTRY, POST_INIT_HOOKS,
PRE_TRAIN_STEP_HOOKS and their register_* decorators into registry.py.
This breaks the circular dependency between passes.py and memory_policy.py.
